Decrease RenderScript Target API to 18
Because it's recommended to "set this value to the lowest
API level able to provide all the functionality" (see
https://developer.android.com/guide/topics/renderscript/compute).

Targeting (RenderScript!) API 24 seems to make Google Play
not transfer the support libraries to devices running Nougat
or higher. If you're installing the (release) APK manually
everything works just fine. Install the APK with Google Play
and the app will crash.