Remove hex detection for SSID
Unfortunately, we can never know whether a string consisting only
of 0-f is a hex string. So we just make this one assumption that
a string of 0-f that is exactly 64 characters long, is _probably_
a hex string for raw PSK. Of course, this may fail for any string
that just happens to be 64 characters long and contains just 0-f.