Remove test for hex strings in WiFi codes
Because SSID/password are no longer quoted in the HashMap.

Instead, both will be quoted when they're used in a WifiConfiguration
since Android Q's WifiNetworkSuggestion.Builder never needs those
values quoted.

Also, since it is fundamentally impossible to tell whether a string
is a hex representation or happens to just consist of characters
from 0 to f, the quotes are now only omitted for a password string
that is exactly 64 characters long and contains only 0-f (raw PSK).