Remove DatabaseRepository and export binary as hex
Well, it was nice seeing Flow in action but I think it's slowly
becoming more of a burden than a benefit.

The traditional Cursor approach sure isn't as nice, but it's much
more simple, does use less resources (because we never need to put
things into an object) and it's less code too.

Also, scans that contain binary data are now always exported as
a hex string.