Replace static inset callback for view callbacks
Using a common static callback function to set the paddings from
WindowInsets creates memory leaks because the function holds
references to the views and so to the Activity or Fragment.

Every view needs to have its own WindowInsetsListener that cannot
hold any references from its Activity or Fragment.

So I'm finally biting the bullet and do it like Chris Banes
described here:
https://medium.com/androiddevelopers/windowinsets-listeners-to-layouts-8f9ccc8fa4d1